Added token_sort_ratio and ratio
Added token_sort_ratio and ratio scores together to see if we can more accurately find the character the user is looking for.

Previously: when search for "Karin (Alchemist)", despite having the exact name, when trying to find the unit Karin, Christna would have a higher score compared to Karin when looking for units, so it would always show Christna.